AI: Extract circuit breaker fom threads.
