+++ trunk/src/app/srs_app_rtc_conn.cpp
@@ -19,6 +19,7 @@ using namespace std;
 #include <queue>
 #include <sstream>
 
+#include <srs_app_circuit_breaker.hpp>
 #include <srs_app_config.hpp>
 #include <srs_app_http_api.hpp>
 #include <srs_app_http_hooks.hpp>
