+++ trunk/src/app/srs_app_rtsp_source.cpp
@@ -6,6 +6,7 @@
 
 #include <srs_app_rtsp_source.hpp>
 
+#include <srs_app_circuit_breaker.hpp>
 #include <srs_app_config.hpp>
 #include <srs_app_hybrid.hpp>
 #include <srs_app_rtc_source.hpp>
